pyproject.toml
[project]
name = "aidops"
version = "0.1.0"
description = "Humanitarian assessment profiles and operational tooling, built on PublicSchema"
requires-python = ">=3.12"
license = "Apache-2.0"
dependencies = [
    "publicschema-build",
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "openpyxl>=3.1",
    "babel>=2.14",
    "rdflib>=7.6.0",
]

[tool.uv]
dev-dependencies = [
    "publicschema-build",
    "pyld>=3.0.0",
    "pyshacl>=0.31.0",
    "pytest>=8.0",
]

[tool.uv.sources]
publicschema-build = { path = "../../../apps/publicschema-build", editable = true }

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[project.scripts]
ao-build = "build.cli:main"
